Update NibbleCollector to increase readout buffer size Now allocates ~93MB on startup... (was 0.3MB)

+++ RecoLuminosity/HLXReadOut/HLXCoreLibs/include/NibbleCollector.hh
@@ -125,8 +125,8 @@ namespace HCAL_HLX
     u32 mNumHLXs;
 
     // Circular buffer pointers
-    u8 mWriteBufferPointer;
-    u8 mReadBufferPointer;
+    u16 mWriteBufferPointer;
+    u16 mReadBufferPointer;
 
     // Circular buffers
     HLX_CB_TYPE *circularBuffer;
